requirements.txt
# 基础依赖
numpy>=1.19.0
opencv-python>=4.5.0
pillow>=8.0.0

# 深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# 人脸识别相关
dlib>=19.22.0
face-recognition>=1.3.0

# 语音处理
SpeechRecognition>=3.8.1
pyttsx3>=2.90
PyAudio>=0.2.11

# AI对话
openai>=0.27.0
aiohttp>=3.8.0
async-timeout>=4.0.0

# 硬件控制
RPi.GPIO>=0.7.0
pigpio>=1.78

# 数据处理
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=0.24.0

# 数据库
sqlite3>=3.35.0

# 音频处理
pygame>=2.0.0
sounddevice>=0.4.0

# 工具库
tqdm>=4.62.0
requests>=2.26.0
python-dateutil>=2.8.2
pytz>=2021.3

# 日志和监控
psutil>=5.8.0
logging>=0.5.1.2

# 测试工具
pytest>=6.2.5
pytest-asyncio>=0.16.0

# 文档工具
sphinx>=4.2.0
sphinx-rtd-theme>=0.5.2

# 开发工具
black>=21.9b0
flake8>=3.9.0
isort>=5.9.0 